requirements.txt
discord.py==2.5.1
aiohttp==3.8.5
attrs==23.1.0
certifi==2025.10.5
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
typing-extensions>=4.14.1,<5
yarl==1.9.2
openai==1.32.0
python-dotenv==1.0.1
pydantic-core==2.41.4
httpx~=0.28.1
dynaconf~=3.2.12